# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI Voice Audit API",
    "description": "\nAI Voice Audit API provides a robust set of endpoints for automating the quality assurance process in call centers using advanced AI models.\n\n## Key Features\n- **Asynchronous Audio Processing**: Upload call recordings for transcription and AI-powered evaluation.\n- **Dynamic LLM Selection**: Configure specific AI models per Line of Business (LOB).\n- **Multi-Tenancy**: Secure data isolation and tenant-specific configurations.\n- **Advanced Analytics**: Real-time dashboards and exportable reports.\n\n## Authentication\nThis API uses **JWT tokens** for secure authentication. \n1. Obtain tokens via the `/api/auth/login` endpoint.\n2. Include the access token in the `Authorization` header for all protected endpoints:\n   `Authorization: Bearer <your_access_token>`\n",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/evaluation/save": {
      "post": {
        "tags": [
          "evaluations"
        ],
        "summary": "Save Manual Evaluation",
        "description": "Save the final score and answers for an evaluation after manual review.",
        "operationId": "save_evaluation_api_evaluation_save_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvaluationUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Unauthorized - Authentication required or invalid token",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Could not validate credentials"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Insufficient permissions",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Insufficient permissions to access this resource"
                }
              }
            }
          },
          "404": {
            "description": "Not Found - Resource does not exist",
            "content": {
              "application/json": {
                "example": {
                  "detail": "User not found"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/process-audio": {
      "post": {
        "tags": [
          "evaluations"
        ],
        "summary": "Process Audio File",
        "description": "Upload an audio recording to be transcribed and analyzed asynchronously by AI.\n\nReturns a `task_id` that can be used to poll for status.\n\n## Example curl:\n```bash\ncurl -X POST \"http://localhost:5000/process-audio\"          -H \"Authorization: Bearer <token>\"          -F \"file=@call.wav\"          -F \"lob_id=1\"\n```",
        "operationId": "process_audio_process_audio_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "lob_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "title": "Lob Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_process_audio_process_audio_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Unauthorized - Authentication required or invalid token",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Could not validate credentials"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Invalid input data",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Missing required field: email"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/task-status/{task_id}": {
      "get": {
        "tags": [
          "evaluations"
        ],
        "summary": "Get Task Status",
        "description": "Check the status of a background audio processing task.\n\nReturns 'SUCCESS', 'FAILURE', or 'PROGRESS'. \nIf successful, includes the evaluation result.",
        "operationId": "get_task_status_task_status__task_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "description": "Not Found - Resource does not exist",
            "content": {
              "application/json": {
                "example": {
                  "detail": "User not found"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/evaluations/{evaluation_id}/assign-lob": {
      "post": {
        "tags": [
          "evaluations"
        ],
        "summary": "Assign LOB to Evaluation",
        "description": "Manually assign or change the Line of Business associated with an evaluation record.",
        "operationId": "assign_lob_api_evaluations__evaluation_id__assign_lob_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "evaluation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Evaluation Id"
            }
          },
          {
            "name": "lob_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Lob Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "description": "Not Found - Resource does not exist",
            "content": {
              "application/json": {
                "example": {
                  "detail": "User not found"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/process-bulk-audio": {
      "post": {
        "tags": [
          "evaluations"
        ],
        "summary": "Process Bulk Audio",
        "description": "Upload multiple audio files for parallel background processing.\n\n## Example curl:\n```bash\ncurl -X POST \"http://localhost:5000/process-bulk-audio\"          -H \"Authorization: Bearer <token>\"          -F \"files=@call1.wav\"          -F \"files=@call2.wav\"          -F \"lob_id=1\"\n```",
        "operationId": "process_bulk_audio_process_bulk_audio_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "lob_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "title": "Lob Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_process_bulk_audio_process_bulk_audio_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "Bad Request - Invalid input data",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Missing required field: email"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Body_process_audio_process_audio_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File",
            "description": "The audio file to process (wav/mp3)"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_process_audio_process_audio_post"
      },
      "Body_process_bulk_audio_process_bulk_audio_post": {
        "properties": {
          "files": {
            "items": {
              "type": "string",
              "format": "binary"
            },
            "type": "array",
            "title": "Files",
            "description": "List of audio files (max 10)"
          }
        },
        "type": "object",
        "required": [
          "files"
        ],
        "title": "Body_process_bulk_audio_process_bulk_audio_post"
      },
      "EvaluationUpdate": {
        "properties": {
          "evaluation_id": {
            "type": "integer",
            "title": "Evaluation Id",
            "description": "ID of the evaluation to update"
          },
          "final_score": {
            "type": "integer",
            "title": "Final Score",
            "description": "The manual score assigned by the QA analyst (0-100)",
            "examples": [
              85
            ]
          },
          "ttch": {
            "type": "integer",
            "title": "Ttch",
            "description": "Total Time to Complete Handling (seconds)",
            "examples": [
              300
            ]
          },
          "final_answers": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Final Answers",
            "description": "The finalized answers for each evaluation criterion",
            "examples": [
              {
                "greeting": "yes",
                "hipaa_verification": "no"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "evaluation_id",
          "final_score",
          "ttch",
          "final_answers"
        ],
        "title": "EvaluationUpdate"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "api/auth/login"
          }
        }
      }
    }
  },
...
}